Fixed a few bugs, but caused a performance regression :(
I just realized that we actually need to keep track of the smallest LSN still in the buffer pool. We shouldn't delete any WAL segments that have records with LSNs after this value! For now, I'm just flushing in commit(), hence the performance regression. We also could flush pages when they get too old. Otherwise, pages that are referenced frequently may not ever be flushed.