Develop (#52)
Changes to the commit heuristics

This commit causes the WAL to `fsync()` on each flush. Calling `fsync()` on the database file after each commit is now optional, and the pager does not flush during commit. It writes back very old pages intermittently. This seems to improve performance a bit, since the WAL `fsync()` happens in the background, while the pager `fsync()` happens in the main thread.