Added code for reading, writing, and finding pointermap pages, modified from SQLite pointermap logic
The pointermap pages will contain back pointers for each page in the database. Eventually, node pages will keep their parent pointers on pointermap pages. The pointermap won't serve any purpose for now, but its functionality will be slowly worked in to the tree. Once it's fully integrated, we should be able to implement a vacuum procedure that works while the database is running. With the current design, a vacuum would just be part of a normal transaction.