Saving work
I've overlooked something that seems obvious in hindsight. When we perform a vacuum, the page IDs contained in the WAL become incorrect. If we roll any of those segments, we will likely crash or corrupt the database. Before the file is truncated, we will need to make sure the WAL and pager have both been flushed, and we can't vacuum with pending updates in the WAL.