Removed support for abort
It makes it more difficult to ensure consistency and adds complexity. I'd like to make the WAL code more robust/clean now that the design is more solidified. Also changed commit() a bit to not return an error if the commit record made it to the WAL. If a different error occurs, like if the WAL couldn't advance, it'll be returned by the next method call. This way we know what is logically in the database.