Cleaned up the WAL reader and recovery routine (#60)
Simplified the WAL. Now, components that want to access the WAL will open a `WalReader` and work directly with the record data. The `WalReader` tries to hide the fact that the WAL is made up of multiple segments. Also removed the "roll backward" functionality. To undo an uncommitted transaction, we just seek to the last commit and read until the end. The full images are disjoint anyway.