Trying out a bit of a different design for the WAL that I'll describe briefly here
Finally going to stop moving to new segments on commit. Instead, just write the commit record like normal and call fsync(). Then just keep on writing records to that segment. During recovery, we may find that the last transaction lacks a commit record, or contains a partial/corrupted record. We will have saved the segment ID and offset of each commit record we encounter, as there may be multiple in one segment (actually, we save the offset right after the commit record, so we can seek to the record following it). Seek the reader back to that position and undo any of its changes that had hit the database disk already (it could be that those changes were still in-memory during the crash). Then, we flush the whole page cache and delete the WAL. We could encounter problems if we fail while deleting WAL segments, but that seems like it would be an unlikely situation. The "commit LSN" would still be correct, however. In repair, we could scan the database and make sure we don't have any changes past that point. If we don't, then we can just finish deleting WAL segments.