Created a LogicalPageId struct that holds the table and page IDs
Pages must be constructed with a logical page ID, which may be used to indicate that the table is unknown. The table ID is not actually used to find the page, it's just for writing WAL records, which is done by the Pager. Keeping the allocate() and acquire() pager methods separate, just indicate LogicalPageId::unknown_page(table_id) when allocating a new page. This allows us to keep the assertion against NULL page IDs in acquire(), which is useful for catching bugs.