Develop
Changed the public API to use a reference where the parameter is not optional, and a pointer where it is optional. The only exception being the case where the argument is a `char *`, which is the case for `Reader::read()` and `Editor::read()`. These methods should take a `std::string &` for their scratch parameters, however, the way that the pager frames are stored (contiguous and aligned to the page size) makes this impossible.