Call fsync on the WAL after every flush, and use a std::multimap for the dirty page table
The extra fsyncs obviously kill performance but they are necessary for ACID compliance. Need to be sure that changes are on the WAL disk before flushing database pages. To help with performance, we could again have flush() and sync() separate, and just update the "flushed_lsn" on sync(), but not on flush(). Working on some unit tests that exercise recovery scenarios where we are actually missing some data that was written after the last fsync.