Less fsync() calls
Only call fsync() on the WAL when the pager needs to write back a page but is unable to due to the flushed_lsn being too low. This lets the WAL flush its tail buffer without having to fsync() (which isn't necessary then anyway). Also, only update the pager recovery_lsn when the database file has had fsync() called on it. Seems to fix the bad performance from the last commit as well.