Got part way through changing to a different WAL implementation
Using something similar to the SQLite WAL format. The DB file is never updated until a checkpoint, which is different than a commit (using the word "commit" now). A commit is just a normal write to the WAL with a special flag set. When commit() is called, all dirty pages are written to the WAL in sequential order. Pages in the WAL within the same commit are overwritten. A checkpoint rolls the WAL and applies changes to the DB. This should be possible by sorting and deduplicating the pages such that we just do a series of sequential writes and 2 fsyncs() on checkpoint.