Develop (#82)
Rewrote the WAL, based on SQLite's implementation. Normally, the pager only writes to the WAL, not the database file. The WAL contents are transferred back to the database during checkpoint operations, which happen (a) after commit if the WAL has reached a certain threshold, or (b) during open/close if necessary. The WAL must be read to get pages that haven't been checkpointed yet.