Concurrency
Sets up CalicoDB to support multithread (each thread must have a separate DB object) and multiprocess concurrency. Honestly, these changes are pretty incomplete, and a lot more validation, testing, and debugging needs to be done. There are just so many changes to the public API that I figured this should be merged now.