Changed the file locking procedure
It seems like we can get by with only shared and exclusive locks on the database file (got rid of reserved and pending modes). Shared memory locking is unchanged. Using compiler intrinsics for some atomic operations, as it was difficult to make the WAL index work with only C++11 atomics.