In the WAL, get rid of the memcpy() calls that can cause data races, and get rid of the attributes that disable TSan. It may be somewhat slower, especially under contention, but this makes everything easier to reason about (and debug). It should also work better with TSan.
Extend the Status object to be a bit more like RocksDB's implementation, which exposes `Code` and `SubCode` enumerators. This makes `Status` twice as large on the stack, unfortunately, but lets us avoid allocations for some statuses (`Status::not_found()` in particular, since this status is returned often from `Table::get()`). It also allows passing additional info that isn't part of the message, which has become necessary.