Develop
Changes the DB::view() and DB::update() methods to be function templates and gets rid of the TxnHandler class. Those methods now take any callable that accepts a Txn& and returns a Status, which is way nicer than having to subclass TxnHandler. Adds subcodes to the Status object to indicate more intricate error conditions. Also, Status no longer stores the code info on the heap, so no allocation is required to create a status with no error message.