Develop (#91)
Add an internal cursor class that tracks the indices and page IDs seen on a root-to-leaf traversal. Use this information to avoid looking up keys that have already been looked up when rebalancing. Later on, this cursor can be used to possibly avoid doing another root-to-leaf search in some situations (for example, when there are multiple consecutive inserts at the end of the key range).