Develop (#95)
When splitting/merging nodes, keep the working set of nodes in an array in the tree object. Release everything that wasn't pushed onto the freelist at the end of each balancing step. This also gets rid of the Page class and simplifies the Node struct to be a wrapper around a NodeHeader, a PageRef *, and a few other things.