Worked on statistics reporting and fixed a bug
Added `stat.h`, which contains a struct for tracking all database statistics. Each statistic is represented by a slot in an array of `U64`s. Add reporting for `fsync()` calls on each file, and the total number of SMOs. Fixed a bug revealed by fuzzing (not setting the "refresh" flag to indicate that the root page was purged if it was dirty during a rollback).