Allow zero-length keys
Removes the check for an empty key in `Tree::put()` and changes some tests to accommodate. Changes `BlockAllocator::release()` to only allow merging of free blocks that are adjacent or separated by 1 byte in external nodes, since external cells may be as small as 2 bytes (1 byte for the key size of 0, 1 for the value size of 0).

NOTE: I'm pretty sure SQLite pads cell headers such that they are always at least 4 bytes. This simplifies logic in various places, because cells can still be used after they are erased (only the first 4 bytes are overwritten, so the payload is still usable). Also, `BlockAllocator::release()` could be more aggressive, merging free blocks when there are 3 or fewer bytes between them. Any group of 3 or fewer unaccounted-for bytes that isn't in the gap space would have to be a fragment. Currently, we can only use internal cells after they are erased.