Allow zero-length keys
Removes the check for an empty key in `Tree::put()` and changes some tests to accommodate. Changes `BlockAllocator::release()` to only allow merging of free blocks that are adjacent or separated by 1 byte in external nodes, since external cells may be as small as 2 bytes (1 byte for the key size of 0, 1 for the value size of 0).