Refactor the tree (#113)
Allows the database to be modified using a cursor. Adds overloads of `Tx::put()` and `Tx::erase()` that take a cursor instead of a bucket. `Tx::put()` leaves the cursor on the record that was just inserted, using the current cursor position as a hint (may avoid a root-to-leaf traversal). `Tx::erase()` erases the record pointed to by the cursor and leaves the cursor on the next record. Also adds some CMake tests.