Develop (#126)
Added a ton of stuff. Got rid of the bucket object. open_bucket() and create_bucket() each return a cursor over the specified bucket. Don't force use of a const pointer for readonly transactions: that was a bad idea. Merge 2 of the internal cursor classes into 1, renaming it CursorImpl. Add a new test framework, stest, which was modified from nasa/fprime. Store active/inactive cursors for each bucket in separate lists in the tree object. Pull the linked list code out into a header.

I think these changes make the API a bit easier to use. The implementation is definitely a lot simpler this way.