Develop (#135)
Split the allocation code up into 2 parts: 1 to define behavior for malloc(0), realloc(ptr, 0), etc., and 1 for additional functionality. For now, implemented a DefaultAllocator that uses std::malloc()/std::realloc()/std::free(), and a DebugAllocator that tracks allocation sizes and can have a callback registered on it (functionality previously in Alloc). Using a mutex to protect calls to allocator methods.