Cleanup (#142)
Add a CMake cache variable for specifying where to print extra messages during testing. Extra messages are nice for sanity checking things that are otherwise difficult to verify, but are annoying because they pollute the output. Defaults to doing nothing and supports writing to stdout, stderr, and a file specified as -DCALICODB_TEST_LOG_FILE=TARGET on the commandline. Work on the tests and start adding comments to wal.h.