Develop (#146)
Change the pager to not create the WAL until the first read-write transaction is started. Go back to mem*() for transferring and comparing the WAL index headers. Set the page count at the time a temporary WAL is created, since it is already known. Prevents division by 0 when db_size() is called before a write occurs.