Canonicalize file paths before storing them (#149)
Otherwise, a custom Env that doesn't interact with the real filesystem, but also implements concurrency, would get confused in some situations. For example, "/tmp/./file" and "/tmp/file" resolve to the same absolute path but do not compare as equal. Use the default allocators for the concurrency tests. The debug allocators cause very bad contention, due to the fact that they write large memory regions with a global mutex locked.