Increase the maximum local cell size (#151)
Changes the maximum local payload size in leaf nodes to be roughly 1/2 of the usable space on a page (used to be roughly 1/4). The root may overflow on the second cell when it is still a leaf node. This is the largest this value can be before we would require more than 1 overflow cell, and more than 1 extra page during rebalancing, which complicates the rebalancing algorithm.