Refactoring and tests (#157)
Changes the default value of `Options::create_if_missing` to false. It must be set to true if the user wishes for `DB::open()` to create a database if it does not already exist. Adds more tests and refactors a bit for coverage. Also fixes a bug in `temp.cpp`: memset() was called on the pointed-to page, rather than the page pointers themselves.