Refactor the internal containers (#158)
Add a Vector container that supports non-trivial objects. Keep Buffer around, since it is able to use realloc() (it only is allowed to contain trivial types, per a static_assert()), and has a somewhat smaller footprint. Use Vector to implement String, so that StringBuilder doesn't have to trim the allocation each time. Rename getters on Buffer (ptr -> data, len -> size) to be consistent with Slice and others.